requirements.txt
wheel
absl-py==1.4.0
ml_collections
numpy==1.23.1
dm-sonnet==2.0.0
#tensorflow==2.12.0
tensorflow_probability==0.20.1
#pyscipopt==4.3.0
graph_nets
#git+git://github.com/deepmind/graph_nets.git@64771dff0d74ca8e77b1f1dcd5a7d26634356d61